Add methods to work with CircleCollection and remove the spatial classes
This commit adds third spatial type, specifically it is collection
of circles. It implements the methods to save new collections, edit
saved collections with change of its location or diameters of circles
in the collection and also implements operation to remove the circle
at the specific position in this collection. Further, this commit
removed the class implements spatial object into individual module
classes in relevant project package.