Edit operations with CircleCollection and add Multipoint object
This commit edit invalid operations with CircleColletion object.
Implemented operations was restricted from the presence of first
circle from the collection. It computed the coordinates from its
points and therefore the functionality was not correct. In this
commit is created the table CircleCollection in the DB, in which
are saved the needed data to perform the relevant operartion.
This commit also adds the next spatial object, specifically it is
Multipoint object, and the basic operation to perform operations with
it. Further this commit reformats all java classes according to styles
by google-java-format.